reorganize signals directory
