+++ lib/src/signals/signals.dart
@@ -1,6 +1,21 @@
 // Copyright (C) 2023 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 
-export 'logic.dart';
+library signals;
+
+import 'dart:async';
+import 'dart:collection';
+
+import 'package:meta/meta.dart';
+import 'package:rohd/rohd.dart';
+import 'package:rohd/src/exceptions/exceptions.dart';
+import 'package:rohd/src/utilities/sanitizer.dart';
+import 'package:rohd/src/utilities/synchronous_propagator.dart';
+
 export 'logic_array.dart';
 export 'logic_structure.dart';
+export 'logic_value_changed.dart';
+
+part 'const.dart';
+part 'logic.dart';
+part 'wire.dart';
